split capability_statement to version specific groups
