+++ lib/us_core_test_kit/custom_groups/capability_statement/conformance_support_test.rb
@@ -1,6 +1,6 @@
 module USCoreTestKit
   class ConformanceSupportTest < Inferno::Test
-    id :us_core_311_conformance_support
+    id :us_core_conformance_support
     title 'FHIR Server supports the conformance interaction'
     description %(
       The conformance 'whole system' interaction provides a method to get the
