+++ lib/us_core_test_kit/custom_groups/capability_statement/fhir_version_test.rb
@@ -1,6 +1,6 @@
 module USCoreTestKit
   class FHIRVersionTest < Inferno::Test
-    id :us_core_311_fhir_version
+    id :us_core_fhir_version
     title 'Server is using FHIR R4'
     description %(
         This test inspects the CapabilityStatement returned by the server to
