+++ lib/us_core_test_kit/custom_groups/capability_statement/profile_support_test.rb
@@ -1,6 +1,6 @@
 module USCoreTestKit
   class ProfileSupportTest < Inferno::Test
-    id :us_core_311_profile_support
+    id :us_core_profile_support
     title 'Capability Statement lists support for required US Core Profiles'
     description %(
       The US Core Implementation Guide states:
@@ -13,45 +13,6 @@ module USCoreTestKit
       ```
     )
     uses_request :capability_statement
-
-    PROFILES = {
-      'AllergyIntolerance' => ['http://hl7.org/fhir/us/core/StructureDefinition/us-core-allergyintolerance'].freeze,
-      'CarePlan' => ['http://hl7.org/fhir/us/core/StructureDefinition/us-core-careplan'].freeze,
-      'CareTeam' => ['http://hl7.org/fhir/us/core/StructureDefinition/us-core-careteam'].freeze,
-      'Condition' => ['http://hl7.org/fhir/us/core/StructureDefinition/us-core-condition'].freeze,
-      'Device' => ['http://hl7.org/fhir/us/core/StructureDefinition/us-core-implantable-device'].freeze,
-      'DiagnosticReport' => [
-        'http://hl7.org/fhir/us/core/StructureDefinition/us-core-diagnosticreport-lab',
-        'http://hl7.org/fhir/us/core/StructureDefinition/us-core-diagnosticreport-note'
-      ].freeze,
-      'DocumentReference' => ['http://hl7.org/fhir/us/core/StructureDefinition/us-core-documentreference'].freeze,
-      'Encounter' => ['http://hl7.org/fhir/us/core/StructureDefinition/us-core-encounter'].freeze,
-      'Goal' => ['http://hl7.org/fhir/us/core/StructureDefinition/us-core-goal'].freeze,
-      'Immunization' => ['http://hl7.org/fhir/us/core/StructureDefinition/us-core-immunization'].freeze,
-      'Location' => ['http://hl7.org/fhir/us/core/StructureDefinition/us-core-location'].freeze,
-      'Medication' => ['http://hl7.org/fhir/us/core/StructureDefinition/us-core-medication'].freeze,
-      'MedicationRequest' => ['http://hl7.org/fhir/us/core/StructureDefinition/us-core-medicationrequest'].freeze,
-      'Observation' => [
-        'http://hl7.org/fhir/us/core/StructureDefinition/us-core-observation-lab',
-        'http://hl7.org/fhir/us/core/StructureDefinition/pediatric-bmi-for-age',
-        'http://hl7.org/fhir/us/core/StructureDefinition/pediatric-weight-for-height',
-        'http://hl7.org/fhir/us/core/StructureDefinition/us-core-pulse-oximetry',
-        'http://hl7.org/fhir/us/core/StructureDefinition/us-core-smokingstatus',
-        'http://hl7.org/fhir/StructureDefinition/bp',
-        'http://hl7.org/fhir/StructureDefinition/bodyheight',
-        'http://hl7.org/fhir/StructureDefinition/bodyweight',
-        'http://hl7.org/fhir/StructureDefinition/heartrate',
-        'http://hl7.org/fhir/StructureDefinition/resprate',
-        'http://hl7.org/fhir/StructureDefinition/bodytemp',
-        'http://hl7.org/fhir/us/core/StructureDefinition/head-occipital-frontal-circumference-percentile'
-      ].freeze,
-      'Organization' => ['http://hl7.org/fhir/us/core/StructureDefinition/us-core-organization'].freeze,
-      'Patient' => ['http://hl7.org/fhir/us/core/StructureDefinition/us-core-patient'].freeze,
-      'Practitioner' => ['http://hl7.org/fhir/us/core/StructureDefinition/us-core-practitioner'].freeze,
-      'PractitionerRole' => ['http://hl7.org/fhir/us/core/StructureDefinition/us-core-practitionerrole'].freeze,
-      'Procedure' => ['http://hl7.org/fhir/us/core/StructureDefinition/us-core-procedure'].freeze,
-      'Provenance' => ['http://hl7.org/fhir/us/core/StructureDefinition/us-core-provenance'].freeze
-    }.freeze
 
     run do
       assert_resource_type(:capability_statement)
@@ -71,11 +32,13 @@ module USCoreTestKit
 
       assert supported_resources.include?('Patient'), 'US Core Patient profile not supported'
 
-      other_resources = PROFILES.keys.reject { |resource_type| resource_type == 'Patient' }
+      target_profiles = config.options[:target_profiles]
+
+      other_resources = target_profiles.keys.reject { |resource_type| resource_type == 'Patient' }
       other_resources_supported = other_resources.any? { |resource| supported_resources.include? resource }
       assert other_resources_supported, 'No US Core resources other than Patient are supported'
 
-      PROFILES.each do |resource_type, profiles|
+      target_profiles.each do |resource_type, profiles|
         next unless supported_resources.include? resource_type
 
         profiles.each do |profile|
