+++ lib/us_core_test_kit/generator/templates/suite.rb.erb
@@ -1,7 +1,7 @@
 require 'inferno/dsl/oauth_credentials'
 require_relative '../../version'
-require_relative '../../custom_groups/capability_statement_group'
-require_relative '<%= clinical_notes_guidance_group %>'
+require_relative '<%= capability_statement_file_name %>'
+require_relative '<%= clinical_notes_guidance_file_name %>'
 require_relative '../../custom_groups/data_absent_reason_group'
 <% group_file_list.each do |file_name| %>require_relative '<%= file_name %>'
 <% end %>
@@ -46,10 +46,10 @@ module USCoreTestKit
         oauth_credentials :smart_credentials
       end
 
-      group from: :us_core_311_capability_statement
+      group from: :<%= capability_statement_group_id %>
   <% group_id_list.each do |id| %>
       group from: :<%= id %><% end %>
-      group from: :<%= clinical_notes_guidance_id %>
+      group from: :<%= clinical_notes_guidance_group_id %>
       group from: :us_core_311_data_absent_reason
     end
   end
